pyproject.toml
[project]
name = "ai-sim"
version = "0.1.0"
description = "AI 驱动的修仙模拟游戏"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
]
